Kaggle-MountainDEW-ishikawa-m1/requirements.txt
absl-py==2.0.0
alembic==1.12.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
black==23.10.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.7.0
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.45.1
gast==0.4.0
gensim==4.2.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.2
h5py==3.10.0
hydra-core==1.3.2
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
joblib==1.3.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
optuna==3.4.0
packaging==23.2
pandas==2.1.2
pathspec==0.11.2
patsy==0.5.3
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
plotly-express==0.4.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
SQLAlchemy==2.0.23
statsmodels==0.14.0
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
texthero==1.0.5
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.15.0
xgboost==2.0.2
